Removed test_failing_integrals.py
Few tests were moved to test_integrals.py

Some remarks:
1) Timeout stuff removed - we can use pytest-timeout plugin.
2) test_sympyissue_3880, test_sympyissue_4525,
   test_sympyissue_4941 - are hangs.
3) test_sympyissue_4326 - solved issue, there is a test.
4) test_sympyissue_1638b - solved, there is a test with a correct value.
5) test_sympyissue_4895* - solved, a duplicate of sympy/sympy#3940.
   There is a test.
6) test_sympyissue_4992 - solved, there is a test.
7) test_sympyissue_4064 - actually issue sympy/sympy#7163, which is
   fixed, there is a test.

Closes sympy/sympy#4511
